requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
langchain
langchain-postgres
langchain-openai
psycopg[binary]
pydantic
jinja2
